Backend/requirements.txt
fastapi==0.129.0
uvicorn==0.34.3
sqlalchemy==2.0.30
pydantic[email]==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
bcrypt==3.2.2
openai==1.91.0
python-multipart==0.0.22
slowapi==0.1.9
psycopg2-binary==2.9.10
email-validator==2.3.0
alembic==1.15.1
